Revert "feat: Create activity instance siren entity (#665)"
This reverts commit 081605d8e9e6773b308caf625da0e8ead8d5f3c1.
