+++ test/activities/instanceCollections/data/CollectedObject.js
