+++ test/activities/instanceCollections/data/InstanceCollection.js
